New rotator using animation

+++ Shared/Fan/Fan View Model.swift
@@ -28,7 +28,7 @@ class FanViewModel: ObservableObject {
     @Published var showTimerIcon = false
     @Published var showDamperWarning = false
     @Published var showInterlockWarning = false
-    @Published var displayedRPM: Int = .zero
+    @Published var displayedRPM: Double = .zero
 
     var chars: FanCharacteristics {
         model.fanCharacteristics.value
@@ -192,7 +192,7 @@ class FanViewModel: ObservableObject {
                 return (chars.speed, levels) }
             .removeDuplicates(by: { ($0.0 == $1.0) && ($0.1 == $1.1) })
             .map {
-                Int( 80.0 * (Double($0.speed) / max( 1.0, Double($0.levels - 1) ) ) )
+                80.0 * (Double($0.speed) / max( 1.0, Double($0.levels - 1) ) )
             }
 //            .print("\r\rrpm pub")
             .assign(to: &$displayedRPM)
